+++ go/pipeline/base_batch.go
@@ -10,6 +10,7 @@ import (
 	"math"
 	"reflect"
 	"strconv"
+	"time"
 
 	"github.com/valkey-io/valkey-glide/go/v2/constants"
 	"github.com/valkey-io/valkey-glide/go/v2/models"
@@ -1684,7 +1685,7 @@ func (b *BaseBatch[T]) LInsert(key string, insertPosition constants.InsertPositi
 // Parameters:
 //
 //	keys        - The keys of the lists to pop from.
-//	timeoutSecs - The number of seconds to wait for a blocking operation to complete. A value of `0` will block indefinitely.
+//	timeout     - The duration to wait for a blocking operation to complete. A value of `0` will block indefinitely.
 //
 // Command Response:
 //
@@ -1694,10 +1695,10 @@ func (b *BaseBatch[T]) LInsert(key string, insertPosition constants.InsertPositi
 //
 // [valkey.io]: https://valkey.io/commands/blpop/
 // [Blocking Commands]: https://github.com/valkey-io/valkey-glide/wiki/General-Concepts#blocking-commands
-func (b *BaseBatch[T]) BLPop(keys []string, timeoutSecs float64) *T {
+func (b *BaseBatch[T]) BLPop(keys []string, timeout time.Duration) *T {
 	return b.addCmdAndConverter(
 		C.BLPop,
-		append(keys, utils.FloatToString(timeoutSecs)),
+		append(keys, utils.FloatToString(timeout.Seconds())),
 		reflect.Slice,
 		true,
 		internal.ConvertArrayOf[string],
@@ -1717,7 +1718,7 @@ func (b *BaseBatch[T]) BLPop(keys []string, timeoutSecs float64) *T {
 // Parameters:
 //
 //	keys        - The keys of the lists to pop from.
-//	timeoutSecs - The number of seconds to wait for a blocking operation to complete. A value of `0` will block indefinitely.
+//	timeout     - The duration to wait for a blocking operation to complete. A value of `0` will block indefinitely.
 //
 // Command Response:
 //
@@ -1727,10 +1728,10 @@ func (b *BaseBatch[T]) BLPop(keys []string, timeoutSecs float64) *T {
 //
 // [valkey.io]: https://valkey.io/commands/brpop/
 // [Blocking Commands]: https://github.com/valkey-io/valkey-glide/wiki/General-Concepts#blocking-commands
-func (b *BaseBatch[T]) BRPop(keys []string, timeoutSecs float64) *T {
+func (b *BaseBatch[T]) BRPop(keys []string, timeout time.Duration) *T {
 	return b.addCmdAndConverter(
 		C.BRPop,
-		append(keys, utils.FloatToString(timeoutSecs)),
+		append(keys, utils.FloatToString(timeout.Seconds())),
 		reflect.Slice,
 		true,
 		internal.ConvertArrayOf[string],
@@ -1869,7 +1870,7 @@ func (b *BaseBatch[T]) LMPopCount(keys []string, listDirection constants.ListDir
 //
 //	keys          - An array of keys to lists.
 //	listDirection - The direction based on which elements are popped from - see [constants.ListDirection].
-//	timeoutSecs   - The number of seconds to wait for a blocking operation to complete. A value of `0` will block indefinitely.
+//	timeout       - The duration to wait for a blocking operation to complete. A value of `0` will block indefinitely.
 //
 // Command Response:
 //
@@ -1878,7 +1879,7 @@ func (b *BaseBatch[T]) LMPopCount(keys []string, listDirection constants.ListDir
 //
 // [valkey.io]: https://valkey.io/commands/blmpop/
 // [Blocking Commands]: https://github.com/valkey-io/valkey-glide/wiki/General-Concepts#blocking-commands
-func (b *BaseBatch[T]) BLMPop(keys []string, listDirection constants.ListDirection, timeoutSecs float64) *T {
+func (b *BaseBatch[T]) BLMPop(keys []string, listDirection constants.ListDirection, timeout time.Duration) *T {
 	listDirectionStr, err := listDirection.ToString()
 	if err != nil {
 		return b.addError("BLMPop", err)
@@ -1891,7 +1892,7 @@ func (b *BaseBatch[T]) BLMPop(keys []string, listDirection constants.ListDirecti
 
 	// args slice will have 3 more arguments with the keys provided.
 	args := make([]string, 0, len(keys)+3)
-	args = append(args, utils.FloatToString(timeoutSecs), strconv.Itoa(len(keys)))
+	args = append(args, utils.FloatToString(timeout.Seconds()), strconv.Itoa(len(keys)))
 	args = append(args, keys...)
 	args = append(args, listDirectionStr)
 	return b.addCmdAndConverter(C.BLMPop, args, reflect.Map, true, internal.ConvertLMPopResponse)
@@ -1915,7 +1916,7 @@ func (b *BaseBatch[T]) BLMPop(keys []string, listDirection constants.ListDirecti
 //	keys          - An array of keys to lists.
 //	listDirection - The direction based on which elements are popped from - see [constants.ListDirection].
 //	count         - The maximum number of popped elements.
-//	timeoutSecs   - The number of seconds to wait for a blocking operation to complete. A value of `0` will block indefinitely.
+//	timeout       - The duration to wait for a blocking operation to complete. A value of `0` will block indefinitely.
 //
 // Command Response:
 //
@@ -1924,7 +1925,12 @@ func (b *BaseBatch[T]) BLMPop(keys []string, listDirection constants.ListDirecti
 //
 // [valkey.io]: https://valkey.io/commands/blmpop/
 // [Blocking Commands]: https://github.com/valkey-io/valkey-glide/wiki/General-Concepts#blocking-commands
-func (b *BaseBatch[T]) BLMPopCount(keys []string, listDirection constants.ListDirection, count int64, timeoutSecs float64) *T {
+func (b *BaseBatch[T]) BLMPopCount(
+	keys []string,
+	listDirection constants.ListDirection,
+	count int64,
+	timeout time.Duration,
+) *T {
 	listDirectionStr, err := listDirection.ToString()
 	if err != nil {
 		return b.addError("BLMPopCount", err)
@@ -1937,7 +1943,7 @@ func (b *BaseBatch[T]) BLMPopCount(keys []string, listDirection constants.ListDi
 
 	// args slice will have 5 more arguments with the keys provided.
 	args := make([]string, 0, len(keys)+5)
-	args = append(args, utils.FloatToString(timeoutSecs), strconv.Itoa(len(keys)))
+	args = append(args, utils.FloatToString(timeout.Seconds()), strconv.Itoa(len(keys)))
 	args = append(args, keys...)
 	args = append(args, listDirectionStr, constants.CountKeyword, utils.IntToString(count))
 	return b.addCmdAndConverter(C.BLMPop, args, reflect.Map, true, internal.ConvertLMPopResponse)
@@ -2020,7 +2026,7 @@ func (b *BaseBatch[T]) LMove(
 //	destination - The key to the destination list.
 //	wherefrom   - The ListDirection the element should be removed from.
 //	whereto     - The ListDirection the element should be added to.
-//	timeoutSecs - The number of seconds to wait for a blocking operation to complete. A value of `0` will block indefinitely.
+//	timeout     - The duration to wait for a blocking operation to complete. A value of `0` will block indefinitely.
 //
 // Command Response:
 //
@@ -2033,7 +2039,7 @@ func (b *BaseBatch[T]) BLMove(
 	destination string,
 	whereFrom constants.ListDirection,
 	whereTo constants.ListDirection,
-	timeoutSecs float64,
+	timeout time.Duration,
 ) *T {
 	whereFromStr, err := whereFrom.ToString()
 	if err != nil {
@@ -2045,7 +2051,7 @@ func (b *BaseBatch[T]) BLMove(
 	}
 	return b.addCmdAndTypeChecker(
 		C.BLMove,
-		[]string{source, destination, whereFromStr, whereToStr, utils.FloatToString(timeoutSecs)},
+		[]string{source, destination, whereFromStr, whereToStr, utils.FloatToString(timeout.Seconds())},
 		reflect.String,
 		true,
 	)
@@ -2979,7 +2985,7 @@ func (b *BaseBatch[T]) ZCard(key string) *T {
 // Parameters:
 //
 //	keys - The keys of the sorted sets.
-//	timeout - The number of seconds to wait for a blocking operation to complete. A value of
+//	timeout - The duration to wait for a blocking operation to complete. A value of
 //	  `0` will block indefinitely.
 //
 // Command Response:
@@ -2990,10 +2996,10 @@ func (b *BaseBatch[T]) ZCard(key string) *T {
 // [valkey.io]: https://valkey.io/commands/bzpopmin/
 //
 // [Blocking commands]: https://github.com/valkey-io/valkey-glide/wiki/General-Concepts#blocking-commands
-func (b *BaseBatch[T]) BZPopMin(keys []string, timeoutSecs float64) *T {
+func (b *BaseBatch[T]) BZPopMin(keys []string, timeout time.Duration) *T {
 	return b.addCmdAndConverter(
 		C.BZPopMin,
-		append(keys, utils.FloatToString(timeoutSecs)),
+		append(keys, utils.FloatToString(timeout.Seconds())),
 		reflect.Slice,
 		true,
 		internal.ConvertKeyWithMemberAndScore,
@@ -3019,7 +3025,7 @@ func (b *BaseBatch[T]) BZPopMin(keys []string, timeoutSecs float64) *T {
 //	keys          - An array of keys to lists.
 //	scoreFilter   - The element pop criteria - either [options.MIN] or [options.MAX] to pop members with the lowest/highest
 //					scores accordingly.
-//	timeoutSecs   - The number of seconds to wait for a blocking operation to complete. A value of `0` will block
+//	timeout       - The duration to wait for a blocking operation to complete. A value of `0` will block
 //					indefinitely.
 //
 // Command Response:
@@ -3031,7 +3037,7 @@ func (b *BaseBatch[T]) BZPopMin(keys []string, timeoutSecs float64) *T {
 //
 // [valkey.io]: https://valkey.io/commands/bzmpop/
 // [Blocking Commands]: https://github.com/valkey-io/valkey-glide/wiki/General-Concepts#blocking-commands
-func (b *BaseBatch[T]) BZMPop(keys []string, scoreFilter constants.ScoreFilter, timeoutSecs float64) *T {
+func (b *BaseBatch[T]) BZMPop(keys []string, scoreFilter constants.ScoreFilter, timeout time.Duration) *T {
 	scoreFilterStr, err := scoreFilter.ToString()
 	if err != nil {
 		return b.addError("BZMPop", err)
@@ -3046,7 +3052,7 @@ func (b *BaseBatch[T]) BZMPop(keys []string, scoreFilter constants.ScoreFilter, 
 
 	// args slice will have 3 more arguments with the keys provided.
 	args := make([]string, 0, len(keys)+3)
-	args = append(args, utils.FloatToString(timeoutSecs), strconv.Itoa(len(keys)))
+	args = append(args, utils.FloatToString(timeout.Seconds()), strconv.Itoa(len(keys)))
 	args = append(args, keys...)
 	args = append(args, scoreFilterStr)
 	return b.addCmdAndConverter(C.BZMPop, args, reflect.Slice, true, internal.ConvertKeyWithArrayOfMembersAndScores)
@@ -3071,7 +3077,7 @@ func (b *BaseBatch[T]) BZMPop(keys []string, scoreFilter constants.ScoreFilter, 
 //	keys          - An array of keys to lists.
 //	scoreFilter   - The element pop criteria - either [options.MIN] or [options.MAX] to pop members with the lowest/highest
 //					scores accordingly.
-//	timeoutSecs   - The number of seconds to wait for a blocking operation to complete. A value of `0` will block indefinitely.
+//	timeout       - The duration to wait for a blocking operation to complete. A value of `0` will block indefinitely.
 //	opts          - Pop options, see [options.ZMPopOptions].
 //
 // Command Response:
@@ -3086,7 +3092,7 @@ func (b *BaseBatch[T]) BZMPop(keys []string, scoreFilter constants.ScoreFilter, 
 func (b *BaseBatch[T]) BZMPopWithOptions(
 	keys []string,
 	scoreFilter constants.ScoreFilter,
-	timeoutSecs float64,
+	timeout time.Duration,
 	opts options.ZMPopOptions,
 ) *T {
 	scoreFilterStr, err := scoreFilter.ToString()
@@ -3103,7 +3109,7 @@ func (b *BaseBatch[T]) BZMPopWithOptions(
 
 	// args slice will have 5 more arguments with the keys provided.
 	args := make([]string, 0, len(keys)+5)
-	args = append(args, utils.FloatToString(timeoutSecs), strconv.Itoa(len(keys)))
+	args = append(args, utils.FloatToString(timeout.Seconds()), strconv.Itoa(len(keys)))
 	args = append(args, keys...)
 	args = append(args, scoreFilterStr)
 	optionArgs, err := opts.ToArgs()
@@ -5494,7 +5500,7 @@ func (b *BaseBatch[T]) ZLexCount(key string, rangeQuery options.RangeByLex) *T {
 // Parameters:
 //
 //	keys - An array of keys to check for elements.
-//	timeoutSecs - The maximum number of seconds to block (`0` blocks indefinitely).
+//	timeout - The maximum duration to block (`0` blocks indefinitely).
 //
 // Command Response:
 //
@@ -5503,8 +5509,8 @@ func (b *BaseBatch[T]) ZLexCount(key string, rangeQuery options.RangeByLex) *T {
 //
 // [valkey.io]: https://valkey.io/commands/bzpopmax/
 // [Blocking Commands]: https://github.com/valkey-io/valkey-glide/wiki/General-Concepts#blocking-commands
-func (b *BaseBatch[T]) BZPopMax(keys []string, timeoutSecs float64) *T {
-	args := append(keys, utils.FloatToString(timeoutSecs))
+func (b *BaseBatch[T]) BZPopMax(keys []string, timeout time.Duration) *T {
+	args := append(keys, utils.FloatToString(timeout.Seconds()))
 	return b.addCmdAndConverter(C.BZPopMax, args, reflect.Slice, true, internal.ConvertKeyWithMemberAndScore)
 }